+++ src/labbench/testing/store_backend.py
@@ -125,6 +125,8 @@ class TestStore:
 
 @key_adapter()
 class StoreTestDevice(Device):
+    resource: str = attr.value.str(None, kw_only=False, help='address')
+
     def open(self):
         self.backend = TestStore()
         attr.observe(self, self.backend.notification_handler)